fixed a bug where the program breaks when the ant hits a bullet

+++ Game/Bullet.pde
@@ -2,7 +2,7 @@ public class Bullet extends Sprite{
   final float velocity = 10; //placeholder
   int damage;
   
-  public Bullet(String img, int xcoor, int ycoor, int Width, int Height, int dmg){
+  public Bullet(String img, float xcoor, float ycoor, int Width, int Height, int dmg){
     super(img, xcoor, ycoor, Width, Height);
     damage = dmg;
   }
