+++ Game/Plant.pde
@@ -1,6 +1,6 @@
 public class Plant extends Sprite{
   int health, cost;
-  Plant(String img, int xcoor, int ycoor, int Width, int Height, int health, int cost){
+  Plant(String img, float xcoor, float ycoor, int Width, int Height, int health, int cost){
     super(img, xcoor, ycoor, Width, Height);
     this.health = health;
     this.cost = cost;
